Moved VersionMigrationService to a new package and extracted the VersionMigrationService.Migration interface as VersionMigration
