+++ src/main/java/ch/jalu/configme/migration/version/VersionMigration.java
@@ -1,0 +1,19 @@
+package ch.jalu.configme.migration.version;
+
+import ch.jalu.configme.migration.Migration;
+
+/**
+ * @author gamerover98
+ */
+public interface VersionMigration extends Migration {
+
+    /**
+     * @return The current version value (such as 1).
+     */
+    int fromVersion();
+
+    /**
+     * @return The next version value (such as 2).
+     */
+    int toVersion();
+}
